Move CLI functions to separate module outside of main
